Linux: Automatically add Bluez entry in psmovepair
This should make the pairing experience very smooth (at least
on Ubuntu-based systems, for others we might have to add more
code to start/stop bluetoothd).

On Linux, "psmovepair" now requires root privileges to stop
bluetoothd and to write the corresponding files. The advantage
is that pairing is automatic: Pair via USB, press the PS button
and the controller is ready to go.