Automated Pairing for Mac OS X 10.8
This has been tested by me under Mac OS X 10.8, and should
work with older versions of Mac OS X as well.

The current implementation is very defensive (i.e. we wait
for "blued" to shutdown by itself instead of killing it),
which means that pairing a controller that's not known to
the system takes ~ 42 seconds (but it seems very reliable).