Compile everything as position-independent code
This fixes compilation issues where the shared Python
module gets linked to the static PS Move API library.

The solution is not perfect, but for now we opt for
having a nicer installation experience (self-contained
Python and Java modules) instead of doing it in a clean
way (shared libraries all over the place).