Automatic dimming factor during blinking calibration
When using the blinking calibration on the first controller, the
saturation value is taken into account to determine the dimming
factor. This is useful in situations with higher exposure values.

The next controllers will then use the same dimming factor as the
first controller for now (we could think about maybe doing
controller-specific dimming factors to better account for different
colors). If the tracker dimming factor is manually set to 0, this
forces a re-calculation of the dimming factor when the next controller
is added.