Add support for debug symbol packages, Qt.
This would be better as two patches, but I'm not sure if it's worth
the overhead to separate them.

Note that I added a Recommends: for libpsmoveapiqt3.  I'm not sure if
that is ideal, but I didn't want to make a separate -dev package just
for libpsmoveapiqt3, and Requires: was a bit limiting.