Fix inline function definition
Compiling the library according to the C99 standard (which, for instance, is clang's default setting) would fail due to an undefined reference to "psmove_decode_16bit" when linking against libpsmoveapi. This fix makes sure that psmove_decode_16bit() is always resolved in the single translation unit that uses it, while keeping the "inline" hint.