Fix temperature reporting
This reverts the changes made in commit e2bc26e and takes care of
proper conversion to degree Celsius, too.

Previous interpretations were based on wrong assumptions regarding the
source of the raw temperature data in the Input Report. The temperature
is not read from the AK8974 magnetometer, even though that IC is
capable of supplying this info. Instead, an ADC reads the voltage
across a thermistor which sits near the accelerometer. It is the raw
ADC output that ends up in the Input Report.