Add support for interactive pairing on Windows
Establishing a connection between the Move and a Bluetooth radio on
Windows is rather tricky (without the use of custom drivers). The
procedure implemented here is as follows:

  1. configure the Bluetooth radio
  2. find a Move with the given Bluetooth device address
  3. try to enable the HID service for this device
  4. check if that succeeded (i.e. device connected and service enabled)

Since this rarely succeeds on the first try, the whole procedure is
repeated over and over again, requiring the user only to press the
controller's PS button every time the small status LED goes off.

This procedure is known to work on Windows XP and Windows 7. It does not
seem to work on Windows 8.