Update Windows build to MinGW-w64 4.9.1
Since different versions of MinGW-w64 differ in support of clock-related
functions, we do now suggest using version 4.9.1 which was tested on
several Windows systems.

This fix removes the now obsolete redefinitions of unsupported clock
features and gives detailed instructions in the README for fetching the
correct version of MinGW for Windows builds.