Automatically detect BlueZ 5.x
Starting with BlueZ 5, pairing on Linux requires different code than
previous versions. While this code is already present, it currently
requires the user to manually define PSMOVE_BLUEZ5_SUPPORT in the build
in order to be used.

This fix implements automatic detection whether BlueZ version 5.x is used
in the build and handles the required define accordingly.