Fix unidentified issue
I tried the library on a new machine, found two issues with it, and fixed them (this is one of them).  I'm still not sure what the ramifications of the changes are, but they work on the target machine.  The only difference I can identify is that a newer version of Windows 10 has been installed on the new machine.