Handle pairing on Windows 8+ and older versions differently
I do actually have problems when testing with the new timing and
retry modifications on Windows prior to version 8. The old
pairing mechanism seems to work better, so we treat these
versions differently:

No changes for Windows 8 and later. But on older versions the
same procedure as prior to the Windows 8 additions is used.