Remove intrinsics.xml/distortion.xml from settings
Those strings are probably not going to change, and if they do,
we should properly manage the memory of the strings (strdup()
when set, and free() when closing), which makes memory management
of the settings unnecessarily complicated. Plus, it was all relative
to the PS Move API configuration directory, anyway.