Add CMake option PSMOVE_PYTHON_VERSION
This sets a version number in the find_package() call that looks for the
Python lib. Useful if multiple versions of the lib are installed and the
user wants CMake to find a particular one of them.

In its most basic form, this can be used to select between Python 2 and
3 (simply set the option to 2 or 3, respectively). But minor versions
are also supported (e.g. 2.7 vs. 2.6).

Leaving it unset (the default) makes CMake behave the same way as
before.