Actually use OpenCV on macOS builds
The build script would check out a patch for OpenCV 3 and ignore the
previously cloned branch '2.4'. Our build would then search for OpenCV 2
and not find a suitable version.

The patch in question has been backported to OpenCV 2.4 years ago. So we
should be good with simply using that branch without applying any
patches manually.