Fix return value of psmove_update_leds()
psmove_update_leds() would always return failure for Bluetooth-connected
controllers on Linux and for all connections on Windows.

The platforms seem to handle the details of writing output reports very
differently, so we opt for a less strict test and now flag only the
cases where the write operation fails altogether.

Fixes issue #376.